version message interfaces and rename json schemas
