+++ func-complete-check/check-marker/index.ts
@@ -1,7 +1,8 @@
 import { AzureFunction, Context } from "@azure/functions"
+import { MarkCheckMessageV1 } from "../types/message-schemas";
 
-const serviceBusQueueTrigger: AzureFunction = async function(context: Context, mySbMsg: any): Promise<void> {
-    context.log('ServiceBus queue trigger function processed message', mySbMsg);
+const serviceBusQueueTrigger: AzureFunction = async function(context: Context, markCheckMessage: MarkCheckMessageV1): Promise<void> {
+  context.log.warn(`check marking not yet implemented`)
 };
 
 export default serviceBusQueueTrigger;
