+++ func-complete-check/messages/complete-check-message-template.json
