+++ func-complete-check/util-msg-gen/index.ts
@@ -2,8 +2,8 @@
 import { AzureFunction, Context, HttpRequest } from "@azure/functions"
 import * as lz from "lz-string"
 import uuid from "uuid/v4"
-import * as checkMessage from "../message-schemas/complete-check-message-template.json"
-import * as largeCompleteCheck from "../message-schemas/large-complete-check.json"
+import * as checkMessage from "../messages/complete-check-message-template.json"
+import * as largeCompleteCheck from "../messages/large-complete-check.json"
 
 const httpTrigger: AzureFunction = function (context: Context, req: HttpRequest): void {
     const message = JSON.parse(JSON.stringify(checkMessage))
